requirements.txt
faiss-cpu==1.7.4
langchain== 0.0.329
openai== 0.28.0
PyPDF2==3.0.1
streamlit==1.26.0
tiktoken==0.4.0
python-dotenv==1.0.0
ctransformers==0.2.25
torch==2.0.1
tokenizers==0.13.3
transformers==4.36.0
pinecone-client==2.2.2
